+++ src/nvim/api/private/helpers.h
@@ -4,7 +4,7 @@
 #include "nvim/api/private/defs.h"
 #include "nvim/decoration.h"
 #include "nvim/ex_eval.h"
-#include "nvim/getchar.h"
+#include "nvim/ex_map.h"
 #include "nvim/lib/kvec.h"
 #include "nvim/memory.h"
 #include "nvim/vim.h"
