+++ src/nvim/digraph.c
@@ -15,6 +15,7 @@
 #include "nvim/ex_cmds2.h"
 #include "nvim/ex_docmd.h"
 #include "nvim/ex_getln.h"
+#include "nvim/ex_map.h"
 #include "nvim/garray.h"
 #include "nvim/getchar.h"
 #include "nvim/mbyte.h"
