+++ src/nvim/eval/funcs.c
@@ -26,9 +26,11 @@
 #include "nvim/ex_cmds2.h"
 #include "nvim/ex_docmd.h"
 #include "nvim/ex_getln.h"
+#include "nvim/ex_map.h"
 #include "nvim/file_search.h"
 #include "nvim/fileio.h"
 #include "nvim/fold.h"
+#include "nvim/getchar.h"
 #include "nvim/globals.h"
 #include "nvim/if_cscope.h"
 #include "nvim/indent.h"
@@ -5976,7 +5978,6 @@ static void f_localtime(typval_T *argvars, typval_T *rettv, FunPtr fptr)
 {
   rettv->vval.v_number = (varnumber_T)time(NULL);
 }
-
 
 static void get_maparg(typval_T *argvars, typval_T *rettv, int exact)
 {
