+++ src/nvim/ex_docmd.c
@@ -29,6 +29,7 @@
 #include "nvim/ex_docmd.h"
 #include "nvim/ex_eval.h"
 #include "nvim/ex_getln.h"
+#include "nvim/ex_map.h"
 #include "nvim/ex_session.h"
 #include "nvim/file_search.h"
 #include "nvim/fileio.h"
@@ -4890,55 +4891,6 @@ theend:
   return tab_number;
 }
 
-/*
- * ":abbreviate" and friends.
- */
-static void ex_abbreviate(exarg_T *eap)
-{
-  do_exmap(eap, TRUE);          // almost the same as mapping
-}
-
-/*
- * ":map" and friends.
- */
-static void ex_map(exarg_T *eap)
-{
-  /*
-   * If we are sourcing .exrc or .vimrc in current directory we
-   * print the mappings for security reasons.
-   */
-  if (secure) {
-    secure = 2;
-    msg_outtrans(eap->cmd);
-    msg_putchar('\n');
-  }
-  do_exmap(eap, FALSE);
-}
-
-/*
- * ":unmap" and friends.
- */
-static void ex_unmap(exarg_T *eap)
-{
-  do_exmap(eap, FALSE);
-}
-
-/*
- * ":mapclear" and friends.
- */
-static void ex_mapclear(exarg_T *eap)
-{
-  map_clear_mode(eap->cmd, eap->arg, eap->forceit, false);
-}
-
-/*
- * ":abclear" and friends.
- */
-static void ex_abclear(exarg_T *eap)
-{
-  map_clear_mode(eap->cmd, eap->arg, true, true);
-}
-
 static void ex_autocmd(exarg_T *eap)
 {
   // Disallow autocommands from .exrc and .vimrc in current
@@ -7937,25 +7889,6 @@ void do_sleep(long msec)
   // input buffer, otherwise a following call to input() fails.
   if (got_int) {
     (void)vpeekc();
-  }
-}
-
-static void do_exmap(exarg_T *eap, int isabbrev)
-{
-  int mode;
-  char_u *cmdp;
-
-  cmdp = eap->cmd;
-  mode = get_map_mode(&cmdp, eap->forceit || isabbrev);
-
-  switch (do_map((*cmdp == 'n') ? 2 : (*cmdp == 'u'),
-                 eap->arg, mode, isabbrev)) {
-  case 1:
-    emsg(_(e_invarg));
-    break;
-  case 2:
-    emsg(isabbrev ? _(e_noabbr) : _(e_nomap));
-    break;
   }
 }
